+++ webknossos/webknossos/cli/upload.py
@@ -7,7 +7,7 @@ import typer
 from ..client import webknossos_context
 from ..client._defaults import DEFAULT_WEBKNOSSOS_URL
 from ..client._upload_dataset import DEFAULT_SIMULTANEOUS_UPLOADS
-from ..dataset import Dataset
+from ..dataset import Dataset, RemoteFolder
 from ._utils import parse_path
 
 
@@ -42,7 +42,14 @@ def main(
         str | None,
         typer.Option(
             help="Alternative name to rename your dataset on upload to WEBKNOSSOS. "
-            "(if not provided, current name of dataset is used)",
+            "If not provided, current dataset name is used.",
+        ),
+    ] = None,
+    folder: Annotated[
+        str | None,
+        typer.Option(
+            help="WEBKNOSSOS dataset folder in which the dataset should be placed. Specify the folder path as a string, separated by `/`. Note that this is about the folders listed in the dataset dashboard, and is independent of the underlying storage location. Example: `Datasets/mySubfolder`. "
+            "If not provided, the root folder is used.",
         ),
     ] = None,
     jobs: Annotated[
@@ -56,6 +63,9 @@ def main(
     """Upload a dataset to a WEBKNOSSOS server."""
 
     with webknossos_context(url=webknossos_url, token=token):
+        folder_id: None | RemoteFolder = None
+        if folder is not None:
+            folder_id = RemoteFolder.get_by_path(folder)
         Dataset.open(dataset_path=source).upload(
-            new_dataset_name=dataset_name, jobs=jobs
+            new_dataset_name=dataset_name, jobs=jobs, folder_id=folder_id
         )